generic + carousel + url button

+++ handlers/postBacks.js
@@ -5,6 +5,8 @@ const Responses = require('../responses/responses.js');
 const WitEntities = require('../wit/entities');
 const TextHandler = require('./text');
 const ContinualResponse = require('./continualResponses.js');
+const Utils = require('../helpers/utils.js');
+const BasicResponses = require('../responses/basic_responses.js');
 
 const responses = new Responses();
 
@@ -14,16 +16,19 @@ module.exports = class PostBackHandler {
     this.continualResponse = new ContinualResponse(profile, cache, undefined);
     this.profile = profile;
     this.postBack = postBack;
+    let builder = postBack.split(":");
+    this.handler = builder.shift();
+    this.args = builder;
     this.cache = cache;
   }
   
   resolve() {
-    let handleFunction = this[this.postBack];
+    let handleFunction = this[this.handler];
     if (handleFunction !== undefined) {
-      return handleFunction.call(this);
+      return handleFunction.apply(this, this.args);
     }
     else {
-      return this.ret("Uhmm, niečo sa pokazilo. My bad.", "invalid postback name: " + this.postBack);
+      return this.ret("Uhmm, niečo sa pokazilo. My bad.", "invalid postback name: " + this.handler);
     }
   }
   
@@ -38,42 +43,43 @@ module.exports = class PostBackHandler {
     })
   }
   
-  button_save_reminder_yes () {
-    let reminder = this.cache.get(this.sender_psid, "reminder");
-    if (reminder === undefined) {
-      return this.ret("Neskoro");
+  button_save_reminder (save) {
+    let shouldSave = save === "true";
+    if (shouldSave) {
+      let reminder = this.cache.get(this.sender_psid, "reminder");
+      if (reminder === undefined) {
+        return this.ret("Neskoro");
+      }
+      ReminderDB.insertNewReminder(reminder);
+      return this.ret("Idem na to!");
+    } else {
+      var response = new Response("text", "Akoby sa nestalo.").next("text", "Ak nie si spokojný s dátumom, odporúčam používať formát [deň]d[mesiac]m.").next("text", "Príklad: 28d9m pisem pisomku z chemie.");
+      return new Promise (resolve => {
+         resolve(response);
+      });
     }
-    ReminderDB.insertNewReminder(reminder);
-    return this.ret("Idem na to!");
-  }
-  button_save_reminder_no () {
-    var response = new Response("text", "Akoby sa nestalo.").next("text", "Ak nie si spokojný s dátumom, odporúčam používať formát [deň]d[mesiac]m.").next("text", "Príklad: 28d9m pisem pisomku z chemie.");
-    return new Promise (resolve => {
-       resolve(response);
-    })
   }
   
-  button_reminder_delete_yes () {
-    ReminderDB.deleteReminder(this.sender_psid);
-    return this.ret("Bam! A sú fuč.");
-  }
-  button_reminder_delete_no () {
-    return this.ret("Ha! Skoro som ich vymazal.");
-  }
-  
-  button_right_name_yes() {
-    let username = this.cache.get(this.sender_psid, "username");
-    let names = username.split(" ");
-    this.profile.fFirstName(names[0]);
-    this.profile.fSecondName(names[1]);
-    UserDB.addUser(this.sender_psid, names[0], names[1]);
-    return this.ret("Tvoje meno je navždy zaznamenané. No nie je to super?");
+  button_reminder_delete (del) {
+    if (del === "true") {
+      ReminderDB.deleteReminder(this.sender_psid);
+      return this.ret("Bam! A sú fuč.");
+    } else {
+      return this.ret("Ha! Skoro som ich vymazal.");
+    }
   }
   
-  
-  // Lunch postbacks
-  button_right_name_no() {
-    return this.ret("Tak teda nič.");
+  button_right_name(isRight) {
+    if (isRight === "true") {
+      let names = this.cache.get(this.sender_psid, "username");
+      this.profile.fFirstName(names[0]);
+      if (names.length > 1) {
+        this.profile.fSecondName(names[names.length - 1]); 
+      }
+      return this.ret("Tvoje meno je navždy zaznamenané. No nie je to super?");
+    } else {
+      return this.ret("Tak teda nič.");
+    }
   }
   
   button_lunch() {    
@@ -91,11 +97,76 @@ module.exports = class PostBackHandler {
     return new TextHandler().simulate("get_lunch", entities, this.profile, this.cache);
   }
   
-  // Get started
+  // Get started and related
   button_get_started() {
     return new Promise(resolve => {
-      resolve(new Response("text", "Tak poďme na to. Som uväznený duch Jura Hronca.").next("text", "... `Super vtip, čo?").next("text", "Budem sa ti snažiť spríjemniť život na GJH.").next("generic", responses.whatCanIDo(this.sender_psid)));
+      resolve(new Response("text", "Som uväznený duch Jura Hronca... Super vtip, že?").next("text", "Budem sa snažiť spríjemniť ti život na GJH a na našej stránke.").next("generic", responses.menu(this.sender_psid)));
     });  
+  }
+  
+  //CONTENT IS TEMPORARY
+  button_about_gjh() {
+    // INFO ABOUT STUDENTS' ACHIEVEMENTS, NOTABLE SCHOOL EVENTS, ..., NOTABLE ALUMNI
+    return new Promise(resolve => {
+      resolve(new Response('text', BasicResponses.getInfo('gjh')));
+    });
+  }
+  
+  button_about_hronec() {
+    return new Promise(resolve => {
+      resolve(new Response('text', BasicResponses.getInfo('hronec')));
+    });
+  }
+  
+  
+  // User identification
+  button_identify_user() {
+    return new Promise(resolve => {
+      resolve(new Response('generic', responses.userTypes('general', this.sender_psid)));
+    });
+  }
+  
+  button_identify_gjh() {
+    return new Promise(resolve => {
+      resolve(new Response('generic', responses.userTypes('gjh', this.sender_psid)));
+    });
+  }
+  
+  button_identify_stranger() {
+    return new Promise(resolve => {
+      resolve(new Response('generic', responses.userTypes('stranger', this.sender_psid)));
+    });
+  }
+  
+  // Modify content based on user
+  button_user_teacher() {
+    return new Promise(resolve => {
+      resolve(new Response('text', 'teacher'));
+    });
+  }
+  
+  button_user_student() {
+    return new Promise(resolve => {
+      resolve(new Response('text', 'student'));
+    });
+  }
+  
+  button_user_parent() {
+    return new Promise(resolve => {
+      resolve(new Response('text', 'parent'));
+    });
+  }
+  
+  button_user_applicant() {
+    return new Promise(resolve => {
+      resolve(new Response('text', 'applicant'));
+    });
+  }
+  
+  button_show_powers() {
+    return new Promise(resolve => {
+      resolve(new Response('generic', responses.mainFunctions(this.sender_psid)));
+    });
   }
   
   set_class() {
